avm-etl/requirements.txt
# GPR AVM ETL - Trestle Historical Data Loader
# Python 3.9+

# HTTP Client
httpx[http2]>=0.27.0

# Database (async PostgreSQL - no psycopg2 needed)
asyncpg>=0.29.0

# Environment & Configuration
python-dotenv>=1.0.0

# Progress & Logging
tqdm>=4.66.0
structlog>=24.0.0

# Date handling
python-dateutil>=2.8.0
